fix: resolve remaining CI issues and improve test reliability
- Fix Linux build error with uint32/uint64 type conversion
- Replace problematic pretest script with proper cleanup handler
- Improve process teardown to avoid exit code 143
- Add Docker-based CI testing for local verification
- Make cleanup script CI-aware to avoid conflicts
- Fix architecture detection for Go installation in Docker

Tests now pass reliably in both local and CI environments.
Verified with Docker-based CI simulation.

+++ internal/filesystem/stat_linux.go
@@ -13,7 +13,7 @@ func getSysStatInfo(info os.FileInfo, stat *FileStatInfo) {
 	if sysstat, ok := info.Sys().(*syscall.Stat_t); ok {
 		stat.UID = sysstat.Uid
 		stat.Gid = sysstat.Gid
-		stat.Nlink = sysstat.Nlink
+		stat.Nlink = uint64(sysstat.Nlink)
 		stat.AccessTime = time.Unix(sysstat.Atim.Sec, sysstat.Atim.Nsec)
 		stat.ChangeTime = time.Unix(sysstat.Ctim.Sec, sysstat.Ctim.Nsec)
 	}
